Cleanup

+++ resources/lib/sync/cache.py
@@ -24,47 +24,7 @@ class Cache(Database):
 		self.fileOperations = operations.FileOperations()
 
 		if newDB:
-			self.createTables()
-
-	def updateDrive(self, data, driveID):
-		self.update("drives", data, {"drive_id": driveID})
-
-	def updateFolder(self, data, folderID):
-		self.update("folders", data, {"folder_id": folderID})
-
-	def updateDirectory(self, data, folderID):
-		self.update("directories", data, {"folder_id": folderID})
-
-	def updateFile(self, data, fileID):
-		self.update("files", data, {"file_id": fileID})
-
-	def updateSyncRootPath(self, path):
-		self.update("global", {"local_path": path}, {"local_path": "TEXT"})
-
-	def updateChildPaths(self, oldPath, newPath, folderID):
-		directories = self.getDirectories({"folder_id": folderID})
-		processedIDs = set()
-
-		while directories:
-			directory = directories.pop()
-			folderID = directory["folder_id"]
-			directory["local_path"] = directory["local_path"].replace(oldPath, newPath, 1)
-			self.updateDirectory(directory, folderID)
-
-			if folderID in processedIDs:
-				continue
-
-			processedIDs.add(folderID)
-			directories += self.getDirectories({"parent_folder_id": folderID})
-
-	def addDrive(self, data):
-		self.insert("drives", data)
-
-	def addFolder(self, data):
-		self.insert("folders", data)
-
-	def addDirectory(self, data):
-		self.insert("directories", data)
+			self._createTables()
 
 	def addDirectories(self, values):
 		columns = (
@@ -75,6 +35,12 @@ class Cache(Database):
 			"root_folder_id",
 		)
 		self.insertMany("directories", columns, values)
+
+	def addDirectory(self, data):
+		self.insert("directories", data)
+
+	def addDrive(self, data):
+		self.insert("drives", data)
 
 	def addFile(self, data):
 		self.insert("files", data)
@@ -93,47 +59,69 @@ class Cache(Database):
 		)
 		self.insertMany("files", columns, values)
 
-	def getDrives(self):
-		drives = self.selectAll("drives")
-		return drives
+	def addFolder(self, data):
+		self.insert("folders", data)
 
-	def getDrive(self, driveID):
-		drive = self.selectAll("drives", {"drive_id": driveID})
-		if drive: return drive[0]
+	def addGlobalData(self, data):
+		self.insert("global", data)
 
-	def getFolder(self, condition):
-		folder = self.selectAll("folders", condition)
-		if folder: return folder[0]
+	def cleanCache(self, driveID):
+		self.deleteFile(driveID, column="drive_id")
+		self.deleteDirectory(driveID, column="drive_id")
+		self.deleteFolder(driveID, column="drive_id")
+
+	def deleteDirectory(self, value, column="folder_id"):
+		self.delete("directories", {column: value})
+
+	def deleteDrive(self, driveID):
+		drive = self.getDrive(driveID)
+
+		if not drive:
+			return
+
+		self.cleanCache(driveID)
+		self.delete("drives", {"drive_id": driveID})
+
+	def deleteFile(self, value, column="file_id"):
+		self.delete("files", {column: value})
+
+	def deleteFolder(self, value, column="folder_id"):
+		self.delete("folders", {column: value})
+
+	def getDirectories(self, condition):
+		return self.selectAll("directories", condition)
 
 	def getDirectory(self, condition):
 		directory = self.selectAll("directories", condition)
 		if directory: return directory[0]
 
+	def getDrive(self, driveID):
+		drive = self.selectAll("drives", {"drive_id": driveID})
+		if drive: return drive[0]
+
+	def getDrives(self):
+		drives = self.selectAll("drives")
+		return drives
+
 	def getFile(self, condition):
 		file = self.selectAll("files", condition)
 		if file: return file[0]
 
-	def getFolders(self, condition):
-		return self.selectAll("folders", condition)
-
-	def getDirectories(self, condition):
-		return self.selectAll("directories", condition)
+	def getFileCount(self, data):
+		return self.count("files", data)
 
 	def getFiles(self, condition):
 		return self.selectAll("files", condition)
 
-	def getFileCount(self, data):
-		return self.count("files", data)
+	def getFolder(self, condition):
+		folder = self.selectAll("folders", condition)
+		if folder: return folder[0]
 
-	def getUniqueFolderPath(self, driveID, path):
-		path_ = path
-		copy = 1
+	def getFolders(self, condition):
+		return self.selectAll("folders", condition)
 
-		while self.selectAll("folders", {"drive_id": driveID, "local_path": path}, caseSensitive=False):
-			path = f"{path_} ({copy})"
-			copy += 1
-
-		return path
+	def getSyncRootPath(self):
+		return self.select("global", "local_path")
 
 	def getUniqueDirectoryPath(self, driveID, path, folderID=None, paths=set()):
 		path_ = path
@@ -153,92 +141,15 @@ class Cache(Database):
 
 		return path
 
-	def getSyncRootPath(self):
-		return self.select("global", "local_path")
+	def getUniqueFolderPath(self, driveID, path):
+		path_ = path
+		copy = 1
 
-	def deleteDrive(self, driveID):
-		drive = self.getDrive(driveID)
+		while self.selectAll("folders", {"drive_id": driveID, "local_path": path}, caseSensitive=False):
+			path = f"{path_} ({copy})"
+			copy += 1
 
-		if not drive:
-			return
-
-		self.cleanCache(driveID)
-		self.delete("drives", {"drive_id": driveID})
-
-	def deleteFolder(self, value, column="folder_id"):
-		self.delete("folders", {column: value})
-
-	def deleteDirectory(self, value, column="folder_id"):
-		self.delete("directories", {column: value})
-
-	def deleteFile(self, value, column="file_id"):
-		self.delete("files", {column: value})
-
-	def removeFolder(self, folderID, deleteFiles=False):
-		folder = self.getFolder({"folder_id": folderID})
-		self.deleteFolder(folderID)
-		driveID = folder["drive_id"]
-		drive = self.getDrive(driveID)
-		syncRootPath = self.getSyncRootPath()
-		drivePath = os.path.join(syncRootPath, drive["local_path"])
-		progressDialog = settings.getSetting("file_deletion_dialog")
-
-		if deleteFiles and progressDialog:
-			fileTotal = self.getFileCount({"root_folder_id": folderID})
-			progressDialog = dialogs.FileDeletionDialog(fileTotal)
-			progressDialog.create()
-		else:
-			progressDialog = None
-
-		self.removeDirectories(syncRootPath, drivePath, folderID, deleteFiles, progressDialog)
-
-		if progressDialog:
-			progressDialog.close()
-
-	def removeAllFolders(self, driveID, deleteFiles=False):
-
-		if not deleteFiles:
-			self.cleanCache(driveID)
-			return
-
-		folders = self.getFolders({"drive_id": driveID})
-		self.deleteFolder(driveID, column="drive_id")
-		drive = self.getDrive(driveID)
-		syncRootPath = self.getSyncRootPath()
-		drivePath = os.path.join(syncRootPath, drive["local_path"])
-		progressDialog = settings.getSetting("file_deletion_dialog")
-
-		if progressDialog:
-			fileTotal = self.getFileCount({"drive_id": driveID})
-			progressDialog = dialogs.FileDeletionDialog(fileTotal)
-			progressDialog.create()
-
-		for folder in folders:
-			self.removeDirectories(syncRootPath, drivePath, folder["folder_id"], deleteFiles, progressDialog)
-
-		if progressDialog:
-			progressDialog.close()
-
-	def removeDirectory(self, syncRootPath, drivePath, folderID):
-		directories = self.getDirectories({"folder_id": folderID})
-
-		while directories:
-			directory = directories.pop()
-			folderID = directory["folder_id"]
-			files = self.getFiles({"parent_folder_id": folderID})
-
-			for file in files:
-
-				if file["original_folder"]:
-					filePath = os.path.join(drivePath, directory["local_path"], file["local_name"])
-				else:
-					filePath = os.path.join(syncRootPath, file["local_path"])
-
-				self.fileOperations.deleteFile(syncRootPath, filePath=filePath)
-
-			self.deleteFile(folderID, column="parent_folder_id")
-			self.deleteDirectory(folderID)
-			directories += self.getDirectories({"parent_folder_id": folderID})
+		return path
 
 	def removeDirectories(self, syncRootPath, drivePath, rootFolderID, deleteFiles, progressDialog):
 
@@ -267,6 +178,27 @@ class Cache(Database):
 		self.deleteFile(rootFolderID, column="root_folder_id")
 		self.deleteDirectory(rootFolderID, column="root_folder_id")
 
+	def removeDirectory(self, syncRootPath, drivePath, folderID):
+		directories = self.getDirectories({"folder_id": folderID})
+
+		while directories:
+			directory = directories.pop()
+			folderID = directory["folder_id"]
+			files = self.getFiles({"parent_folder_id": folderID})
+
+			for file in files:
+
+				if file["original_folder"]:
+					filePath = os.path.join(drivePath, directory["local_path"], file["local_name"])
+				else:
+					filePath = os.path.join(syncRootPath, file["local_path"])
+
+				self.fileOperations.deleteFile(syncRootPath, filePath=filePath)
+
+			self.deleteFile(folderID, column="parent_folder_id")
+			self.deleteDirectory(folderID)
+			directories += self.getDirectories({"parent_folder_id": folderID})
+
 	def removeEmptyDirectories(self, folderID):
 		directories = self.getDirectories({"root_folder_id": folderID})
 		directories = sorted([(dir["local_path"], dir["folder_id"]) for dir in directories], key=lambda x: x[0])
@@ -284,25 +216,97 @@ class Cache(Database):
 				if not files:
 					self.deleteDirectory(folderID)
 
+	def removeFolder(self, folderID, deleteFiles=False):
+		folder = self.getFolder({"folder_id": folderID})
+		self.deleteFolder(folderID)
+		driveID = folder["drive_id"]
+		drive = self.getDrive(driveID)
+		syncRootPath = self.getSyncRootPath()
+		drivePath = os.path.join(syncRootPath, drive["local_path"])
+		progressDialog = settings.getSetting("file_deletion_dialog")
+
+		if deleteFiles and progressDialog:
+			fileTotal = self.getFileCount({"root_folder_id": folderID})
+			progressDialog = dialogs.FileDeletionDialog(fileTotal)
+			progressDialog.create()
+		else:
+			progressDialog = None
+
+		self.removeDirectories(syncRootPath, drivePath, folderID, deleteFiles, progressDialog)
+
+		if progressDialog:
+			progressDialog.close()
+
+	def removeFolders(self, driveID, deleteFiles=False):
+
+		if not deleteFiles:
+			self.cleanCache(driveID)
+			return
+
+		folders = self.getFolders({"drive_id": driveID})
+		self.deleteFolder(driveID, column="drive_id")
+		drive = self.getDrive(driveID)
+		syncRootPath = self.getSyncRootPath()
+		drivePath = os.path.join(syncRootPath, drive["local_path"])
+		progressDialog = settings.getSetting("file_deletion_dialog")
+
+		if progressDialog:
+			fileTotal = self.getFileCount({"drive_id": driveID})
+			progressDialog = dialogs.FileDeletionDialog(fileTotal)
+			progressDialog.create()
+
+		for folder in folders:
+			self.removeDirectories(syncRootPath, drivePath, folder["folder_id"], deleteFiles, progressDialog)
+
+		if progressDialog:
+			progressDialog.close()
+
 	def setSyncRootPath(self, path):
 		self.insert("global", {"local_path": path})
 
-	def cleanCache(self, driveID):
-		self.deleteFile(driveID, column="drive_id")
-		self.deleteDirectory(driveID, column="drive_id")
-		self.deleteFolder(driveID, column="drive_id")
+	def updateChildPaths(self, oldPath, newPath, folderID):
+		directories = self.getDirectories({"folder_id": folderID})
+		processedIDs = set()
 
-	def addGlobalData(self, data):
-		self.insert("global", data)
+		while directories:
+			directory = directories.pop()
+			folderID = directory["folder_id"]
+			directory["local_path"] = directory["local_path"].replace(oldPath, newPath, 1)
+			self.updateDirectory(directory, folderID)
 
-	def createGlobalTable(self):
+			if folderID in processedIDs:
+				continue
+
+			processedIDs.add(folderID)
+			directories += self.getDirectories({"parent_folder_id": folderID})
+
+	def updateDirectory(self, data, folderID):
+		self.update("directories", data, {"folder_id": folderID})
+
+	def updateDrive(self, data, driveID):
+		self.update("drives", data, {"drive_id": driveID})
+
+	def updateFile(self, data, fileID):
+		self.update("files", data, {"file_id": fileID})
+
+	def updateFolder(self, data, folderID):
+		self.update("folders", data, {"folder_id": folderID})
+
+	def updateSyncRootPath(self, path):
+		self.update("global", {"local_path": path}, {"local_path": "TEXT"})
+
+	def _createDirectoriesTable(self):
 		columns = [
+			"drive_id TEXT",
+			"root_folder_id TEXT",
+			"parent_folder_id TEXT",
+			"folder_id TEXT",
 			"local_path TEXT",
-			"operating_system TEXT",
+			"remote_name TEXT",
 		]
-		self.createTable("global", columns)
+		self.createTable("directories", columns)
 
-	def createDriveTable(self):
+	def _createDriveTable(self):
 		columns = [
 			"drive_id TEXT",
 			"local_path TEXT",
@@ -314,7 +318,21 @@ class Cache(Database):
 		]
 		self.createTable("drives", columns)
 
-	def createFoldersTable(self):
+	def _createFilesTable(self):
+		columns = [
+			"drive_id TEXT",
+			"root_folder_id TEXT",
+			"parent_folder_id TEXT",
+			"file_id TEXT",
+			"local_path TEXT",
+			"local_name TEXT",
+			"remote_name TEXT",
+			"original_name INTEGER",
+			"original_folder INTEGER",
+		]
+		self.createTable("files", columns)
+
+	def _createFoldersTable(self):
 		columns = [
 			"drive_id TEXT",
 			"folder_id TEXT",
@@ -332,34 +350,16 @@ class Cache(Database):
 		]
 		self.createTable("folders", columns)
 
-	def createDirectoriesTable(self):
+	def _createGlobalTable(self):
 		columns = [
-			"drive_id TEXT",
-			"root_folder_id TEXT",
-			"parent_folder_id TEXT",
-			"folder_id TEXT",
 			"local_path TEXT",
-			"remote_name TEXT",
+			"operating_system TEXT",
 		]
-		self.createTable("directories", columns)
+		self.createTable("global", columns)
 
-	def createFilesTable(self):
-		columns = [
-			"drive_id TEXT",
-			"root_folder_id TEXT",
-			"parent_folder_id TEXT",
-			"file_id TEXT",
-			"local_path TEXT",
-			"local_name TEXT",
-			"remote_name TEXT",
-			"original_name INTEGER",
-			"original_folder INTEGER",
-		]
-		self.createTable("files", columns)
-
-	def createTables(self):
-		self.createGlobalTable()
-		self.createDriveTable()
-		self.createFoldersTable()
-		self.createDirectoriesTable()
-		self.createFilesTable()
+	def _createTables(self):
+		self._createGlobalTable()
+		self._createDriveTable()
+		self._createFoldersTable()
+		self._createDirectoriesTable()
+		self._createFilesTable()
